src/requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Cython==0.28.1
DistributedLock==1.2
Django==1.11.4
django-annoying==0.10.4
django-m2m-history==0.3.6
django-oauth-tokens==0.6.3
django-picklefield==1.0.0
django-taggit==0.22.2
ez-setup==0.9
future==0.16.0
idna==2.6
numpy==1.14.2
oauthlib==2.0.7
PySocks==1.6.8
pystan==2.17.1.0
python-memcached==1.59
python-twitter==3.4.1
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
tweepy==3.6.0
urllib3==1.22
